refactor(pnpr): split oversized modules
