+++ pnpr/crates/pnpr/src/resolver.rs
@@ -42,27 +42,25 @@
 //! so private dependencies resolve via a pnpr-managed upstream credential or
 //! fail closed.
 
+mod cache;
 pub(crate) mod osv;
 mod protocol;
+mod request_validation;
 mod resolve;
 mod verdict_cache;
+mod wire;
 
 use std::{
     collections::HashMap,
     path::{Path, PathBuf},
     sync::{Arc, LazyLock, Mutex, OnceLock},
-    time::{Duration, Instant},
+    time::Duration,
 };
 
 use crate::{
     config::Config as RegistryConfig,
-    package_name::PackageName,
     policy::Identity,
-    route::{
-        Footprint, RouteClass, RouteContext, RouteHook, sanitize_registry_tarball_url,
-        strip_url_credentials,
-    },
-    upstream::tarball_basename,
+    route::{Footprint, RouteContext, RouteHook},
 };
 
 use axum::{
@@ -72,25 +70,38 @@ use axum::{
 };
 use indexmap::IndexMap;
 use pnpm_config::Config as PacquetConfig;
-use pnpm_lockfile::{
-    Lockfile, LockfileResolution, TarballResolution, TarballRevision, is_git_hosted_tarball_url,
-    pick_registry_for_package,
-};
+use pnpm_lockfile::Lockfile;
 use pnpm_lockfile_verification::{collect_resolution_policy_violations, hash_lockfile};
 use pnpm_network::{AuthHeaders, ThrottledClient, UpstreamRouteHook};
-use pnpm_package_manager::{
-    ResolvedPackageHint, build_resolution_verifiers, tarball_url_and_integrity,
-};
+use pnpm_package_manager::build_resolution_verifiers;
 use pnpm_resolving_npm_resolver::{
     InMemoryPackageMetaCache, ObservedDistStats, PackageMetaCache, observed_dist_stats_sink,
 };
 use pnpm_resolving_resolver_base::{PackageVersionGuard, ResolutionVerifier};
 use pnpm_store_dir::StoreDir;
-use sha2::{Digest, Sha256};
 
 pub(crate) use self::osv::{OsvIndex, format_advisory_ids, load_osv_index};
 
-use self::{protocol::ResolveRequest, verdict_cache::VerdictCache};
+use self::{
+    cache::{CachedResolution, cached_resolution, resolution_cache_key, store_resolution},
+    protocol::ResolveRequest,
+    request_validation::{
+        reject_inline_url_auth, reject_invalid_patch_hashes, reject_invalid_registries,
+        reject_off_allowlist_fetches,
+    },
+    verdict_cache::VerdictCache,
+    wire::{
+        StreamObserver, TarballRouter, done_frame, error_frame, frozen_package_frames,
+        ndjson_frames, ndjson_single_frame, ndjson_stream_response, osv_violations_for_lockfile,
+        verify_done_or_osv_violations, violations_frame,
+    },
+};
+
+#[cfg(test)]
+use self::{
+    cache::MAX_RESOLUTION_CACHE_CANDIDATES_PER_KEY,
+    wire::{is_osv_checkable_resolution, package_frame, tarball_url_version},
+};
 
 /// Per-server engine backing the pnpr install endpoint: it holds the
 /// store, cache, and HTTP client used to resolve a client's project. The
@@ -134,14 +145,6 @@ pub(crate) struct Resolver {
     /// Part 1 uses it only to label each resolve's cache class in the
     /// operator debug log; Part 2 keys private cache entries by it.
     resolution_cache_secret: Arc<[u8]>,
-}
-
-struct CachedResolution {
-    lockfile: Lockfile,
-    inserted: Instant,
-    last_used: Instant,
-    footprint: Footprint,
-    descriptor_digest: Option<String>,
 }
 
 impl Resolver {
@@ -650,730 +653,6 @@ pub(crate) async fn handle_verify_lockfile(
     }
 }
 
-const MAX_RESOLUTION_CACHE_ENTRIES: usize = 1024;
-const MAX_RESOLUTION_CACHE_CANDIDATES_PER_KEY: usize = 8;
-
-fn cached_resolution(
-    cache: &Mutex<HashMap<String, Vec<CachedResolution>>>,
-    ttl: Duration,
-    key: &str,
-    route_context: &RouteContext,
-    identity: &Identity,
-) -> Option<Lockfile> {
-    if ttl.is_zero() {
-        return None;
-    }
-    let mut cache = cache.lock().expect("resolution cache poisoned");
-    let candidates = cache.get_mut(key)?;
-    candidates.retain(|candidate| candidate.inserted.elapsed() <= ttl);
-    let Some((candidate_index, _)) = candidates.iter().enumerate().find(|(_, candidate)| {
-        candidate.footprint.is_public() || candidate.footprint.allows(route_context, identity)
-    }) else {
-        if candidates.is_empty() {
-            cache.remove(key);
-        }
-        return None;
-    };
-    let candidate = &mut candidates[candidate_index];
-    candidate.last_used = Instant::now();
-    Some(candidate.lockfile.clone())
-}
-
-fn store_resolution(
-    cache: &Mutex<HashMap<String, Vec<CachedResolution>>>,
-    ttl: Duration,
-    key: String,
-    footprint: Footprint,
-    secret: &[u8],
-    lockfile: &Lockfile,
-) -> bool {
-    if ttl.is_zero() {
-        return false;
-    }
-    let now = Instant::now();
-    let descriptor_digest = footprint.digest(secret);
-    let candidate = CachedResolution {
-        lockfile: lockfile.clone(),
-        inserted: now,
-        last_used: now,
-        footprint,
-        descriptor_digest,
-    };
-    let mut cache = cache.lock().expect("resolution cache poisoned");
-    prune_expired_resolution_cache(&mut cache, ttl);
-    let candidates = cache.entry(key).or_default();
-    if let Some(existing) =
-        candidates.iter_mut().find(|entry| entry.descriptor_digest == candidate.descriptor_digest)
-    {
-        *existing = candidate;
-        return true;
-    }
-    candidates.push(candidate);
-    enforce_candidate_limit(candidates);
-    while count_resolution_candidates(&cache) > MAX_RESOLUTION_CACHE_ENTRIES {
-        if !evict_lru_resolution_candidate(&mut cache, true) {
-            break;
-        }
-    }
-    true
-}
-
-fn prune_expired_resolution_cache(
-    cache: &mut HashMap<String, Vec<CachedResolution>>,
-    ttl: Duration,
-) {
-    cache.retain(|_, candidates| {
-        candidates.retain(|candidate| candidate.inserted.elapsed() <= ttl);
-        !candidates.is_empty()
-    });
-}
-
-fn enforce_candidate_limit(candidates: &mut Vec<CachedResolution>) {
-    while candidates.len() > MAX_RESOLUTION_CACHE_CANDIDATES_PER_KEY {
-        evict_lru_candidate(candidates, true);
-    }
-}
-
-fn evict_lru_candidate(candidates: &mut Vec<CachedResolution>, private_first: bool) {
-    if private_first
-        && let Some(index) = candidates
-            .iter()
-            .enumerate()
-            .filter(|(_, candidate)| !candidate.footprint.is_public())
-            .min_by_key(|(_, candidate)| candidate.last_used)
-            .map(|(index, _)| index)
-    {
-        candidates.remove(index);
-        return;
-    }
-    if let Some(index) = candidates
-        .iter()
-        .enumerate()
-        .min_by_key(|(_, candidate)| candidate.last_used)
-        .map(|(index, _)| index)
-    {
-        candidates.remove(index);
-    }
-}
-
-fn count_resolution_candidates(cache: &HashMap<String, Vec<CachedResolution>>) -> usize {
-    cache.values().map(Vec::len).sum()
-}
-
-fn evict_lru_resolution_candidate(
-    cache: &mut HashMap<String, Vec<CachedResolution>>,
-    private_first: bool,
-) -> bool {
-    let target = lru_resolution_candidate(cache, private_first)
-        .or_else(|| if private_first { lru_resolution_candidate(cache, false) } else { None });
-    let Some((key, index, _)) = target else {
-        return false;
-    };
-    if let Some(candidates) = cache.get_mut(&key)
-        && index < candidates.len()
-    {
-        candidates.remove(index);
-    }
-    if cache.get(&key).is_some_and(Vec::is_empty) {
-        cache.remove(&key);
-    }
-    true
-}
-
-fn lru_resolution_candidate(
-    cache: &HashMap<String, Vec<CachedResolution>>,
-    private_only: bool,
-) -> Option<(String, usize, Instant)> {
-    cache
-        .iter()
-        .filter_map(|(key, candidates)| {
-            candidates
-                .iter()
-                .enumerate()
-                .filter(|(_, candidate)| !private_only || !candidate.footprint.is_public())
-                .min_by_key(|(_, candidate)| candidate.last_used)
-                .map(|(index, candidate)| (key.clone(), index, candidate.last_used))
-        })
-        .min_by_key(|(_, _, last_used)| *last_used)
-}
-
-fn resolution_cache_key(config: &PacquetConfig, request: &ResolveRequest) -> Option<String> {
-    // A revision refresh is an explicit request for current upstream state.
-    // Serving or storing it in the whole-resolution cache would make the next
-    // refresh repeat an old registry selection until the cache TTL expires.
-    if request.update_patches {
-        return None;
-    }
-    let projects: Vec<serde_json::Value> = request
-        .projects_normalized()
-        .into_iter()
-        .map(|project| {
-            serde_json::json!({
-                "dir": project.dir,
-                "name": project.name,
-                "version": project.version,
-                "dependencies": project.dependencies,
-                "devDependencies": project.dev_dependencies,
-                "optionalDependencies": project.optional_dependencies,
-            })
-        })
-        .collect();
-    let input = serde_json::json!({
-        "registry": &config.registry,
-        "registries": &request.registries,
-        "overrides": &request.overrides,
-        "catalogs": &request.catalogs,
-        "patchedDependencies": &request.patched_dependencies,
-        "packageExtensions": &request.package_extensions,
-        "allowUnusedPatches": request.allow_unused_patches,
-        "autoInstallPeers": config.auto_install_peers,
-        "dedupePeers": config.dedupe_peers,
-        "excludeLinksFromLockfile": config.exclude_links_from_lockfile,
-        "projects": projects,
-        "inputLockfileHash": request.lockfile.as_ref().map(hash_lockfile),
-        "frozenLockfile": request.frozen_lockfile,
-        "preferFrozenLockfile": request.prefer_frozen_lockfile,
-        "ignoreManifestCheck": request.ignore_manifest_check,
-        "trustLockfile": request.trust_lockfile,
-        "resolutionMode": request.resolution_mode,
-        "minimumReleaseAge": request.minimum_release_age,
-        "minimumReleaseAgeExclude": &request.minimum_release_age_exclude,
-        "minimumReleaseAgeIgnoreMissingTime": request.minimum_release_age_ignore_missing_time,
-        "trustPolicy": request.trust_policy,
-        "trustPolicyExclude": &request.trust_policy_exclude,
-        "trustPolicyIgnoreAfter": request.trust_policy_ignore_after,
-    });
-    let bytes = serde_json::to_vec(&input).ok()?;
-    let mut hasher = Sha256::new();
-    hasher.update(bytes);
-    Some(format!("{:x}", hasher.finalize()))
-}
-
-#[derive(Clone)]
-struct TarballRouter {
-    context: Arc<RouteContext>,
-    identity: Identity,
-    public_url: String,
-    /// Per-scope registry map (`scope -> registry URL`, plus the default) used
-    /// to classify a registry-resolved package by its *registry* route rather
-    /// than its `dist.tarball` host. See [`Self::route_registry_url`].
-    registries: HashMap<String, String>,
-}
-
-impl TarballRouter {
-    fn new(
-        context: Arc<RouteContext>,
-        identity: Identity,
-        public_url: String,
-        registries: HashMap<String, String>,
-    ) -> Self {
-        Self { context, identity, public_url, registries }
-    }
-
-    /// Route a registry-resolved package's tarball by the **registry** it came
-    /// from, not its `dist.tarball` URL. A split-domain registry serves the
-    /// tarball from a different host than the packument, so classifying by the
-    /// tarball URL would misread a private package as public and leak its raw
-    /// upstream URL. Classifying by the registry origin keeps a private
-    /// package on its `/~<name>/` endpoint; a public one still emits its real
-    /// (anonymously fetchable) tarball URL for a direct CDN download.
-    fn route_registry_url(&self, package: &str, version: &str, tarball_url: &str) -> String {
-        let registry = pick_registry_for_package(&self.registries, package, None);
-        match self.context.classify(&self.identity, &registry, Some(package)) {
-            // The `dist.tarball` is untrusted upstream metadata, so sanitize it
-            // before emitting/caching: drop inline `user:pass@host` userinfo and
-            // any query/fragment a registry could use to carry a signed-URL
-            // token. A genuinely public tarball is anonymously fetchable, so the
-            // sanitized URL still works.
-            RouteClass::Public => sanitize_registry_tarball_url(tarball_url),
-            RouteClass::Hosted { .. } => pnpr_tarball_url(
-                &self.public_url,
-                package,
-                &tarball_filename(package, version, tarball_url),
-            ),
-            RouteClass::Proxied { alias, .. } => upstream_endpoint_tarball_url(
-                &self.public_url,
-                &alias,
-                package,
-                &tarball_filename(package, version, tarball_url),
-            ),
-        }
-    }
-
-    fn route_lockfile(&self, config: &PacquetConfig, lockfile: &Lockfile) -> Lockfile {
-        let mut routed = lockfile.clone();
-        let Some(packages) = routed.packages.as_mut() else {
-            return routed;
-        };
-        for (package_key, metadata) in packages {
-            if !matches!(
-                metadata.resolution,
-                LockfileResolution::Registry(_) | LockfileResolution::Tarball(_),
-            ) {
-                continue;
-            }
-            // A resolution that pins no integrity keeps its original URL:
-            // routing it through the endpoint would hand the client a
-            // mirrored tarball it has no hash to check.
-            let Ok((tarball_url, Some(integrity))) =
-                tarball_url_and_integrity(&metadata.resolution, package_key, config)
-            else {
-                continue;
-            };
-            if !is_http_tarball_url(&tarball_url) || is_git_hosted_tarball_url(&tarball_url) {
-                continue;
-            }
-            let name = package_key.name.to_string();
-            let version = package_key.suffix.version().to_string();
-            let routed_url = self.route_url(&name, &version, &tarball_url);
-            if routed_url == tarball_url.as_ref() {
-                continue;
-            }
-            metadata.resolution = LockfileResolution::Tarball(TarballResolution {
-                tarball: routed_url,
-                integrity: Some(integrity.clone()),
-                revision: None,
-                git_hosted: None,
-                path: None,
-            });
-        }
-        routed
-    }
-
-    fn verification_lockfile(&self, lockfile: &Lockfile) -> Lockfile {
-        let mut upstream = lockfile.clone();
-        let Some(packages) = upstream.packages.as_mut() else {
-            return upstream;
-        };
-        for metadata in packages.values_mut() {
-            let LockfileResolution::Tarball(resolution) = &mut metadata.resolution else {
-                continue;
-            };
-            if let Some(tarball_url) = self.upstream_endpoint_tarball_url(&resolution.tarball) {
-                resolution.tarball = tarball_url;
-            }
-        }
-        upstream
-    }
-
-    fn route_url(&self, package: &str, version: &str, tarball_url: &str) -> String {
-        match self.context.classify(&self.identity, tarball_url, Some(package)) {
-            // A public route keeps its upstream URL: it was fetched
-            // anonymously, so its tarball is anonymously fetchable and pnpr
-            // never mints a per-tarball gateway URL. Any inline userinfo a
-            // malicious/compromised registry embedded in `dist.tarball` is
-            // stripped first, so pnpr never streams or caches it.
-            RouteClass::Public => strip_url_credentials(tarball_url),
-            RouteClass::Hosted { .. } => pnpr_tarball_url(
-                &self.public_url,
-                package,
-                &tarball_filename(package, version, tarball_url),
-            ),
-            RouteClass::Proxied { alias, .. } => upstream_endpoint_tarball_url(
-                &self.public_url,
-                &alias,
-                package,
-                &tarball_filename(package, version, tarball_url),
-            ),
-        }
-    }
-
-    /// Reverse a `/~<name>/<pkg>/-/<file>` endpoint tarball URL back to its
-    /// upstream URL so an input lockfile carrying endpoint URLs can be verified
-    /// against the real registry. Returns `None` for any other URL, and for an
-    /// endpoint the caller is not authorized for (so verification cannot be
-    /// used as an oracle for an upstream the caller cannot reach).
-    fn upstream_endpoint_tarball_url(&self, tarball_url: &str) -> Option<String> {
-        let prefix = format!("{}/~", self.public_url.trim_end_matches('/'));
-        let route = tarball_url.strip_prefix(&prefix)?;
-        let (upstream, rest) = route.split_once('/')?;
-        let registry = self.context.upstream_registry(&self.identity, upstream)?;
-        Some(format!("{}/{rest}", registry.trim_end_matches('/')))
-    }
-}
-
-fn tarball_filename(package: &str, version: &str, tarball_url: &str) -> String {
-    tarball_basename(tarball_url).map_or_else(
-        || {
-            PackageName::parse(package).map_or_else(
-                |_| format!("{package}-{version}.tgz"),
-                |name| name.tarball_name_for_version(version),
-            )
-        },
-        str::to_string,
-    )
-}
-
-fn pnpr_tarball_url(public_url: &str, package: &str, filename: &str) -> String {
-    format!("{}/{package}/-/{filename}", public_url.trim_end_matches('/'))
-}
-
-/// The `/~<name>/<package>/-/<filename>` registry-endpoint URL a proxied
-/// route's tarball is served through. Canonical for a client whose scope is
-/// configured at `https://<pnpr>/~<name>/`, so the lockfile entry collapses
-/// to integrity-only; the upstream URL and credential stay server-side.
-fn upstream_endpoint_tarball_url(
-    public_url: &str,
-    upstream: &str,
-    package: &str,
-    filename: &str,
-) -> String {
-    format!("{}/~{upstream}/{package}/-/{filename}", public_url.trim_end_matches('/'))
-}
-
-/// NDJSON content type for the `/-/pnpr/v0/resolve` response. One JSON object
-/// per line; the client parses frames as they arrive. Excluded from the
-/// server's gzip [`CompressionLayer`](crate::server) so frames flush to
-/// the client incrementally rather than being buffered by the encoder.
-const NDJSON_CONTENT_TYPE: &str = "application/x-ndjson";
-const PROJECT_TRANSFORMS_HEADER: &str = "pnpr-project-transforms";
-const PROJECT_TRANSFORMS_VERSION: &str = "1";
-
-/// [`ResolutionObserver`](pnpm_package_manager::ResolutionObserver)
-/// that turns each resolved tarball into a `package` NDJSON frame and
-/// pushes it down the response channel. `on_resolved` is best-effort: a
-/// closed channel (client hung up) or a serialization failure drops the
-/// frame silently — the resolve still runs to completion server-side.
-struct StreamObserver {
-    tx: tokio::sync::mpsc::UnboundedSender<Vec<u8>>,
-    package_version_guard: Option<Arc<dyn PackageVersionGuard>>,
-    tarball_router: TarballRouter,
-}
-
-impl pnpm_package_manager::ResolutionObserver for StreamObserver {
-    fn on_resolved(&self, hint: pnpm_package_manager::ResolvedPackageHint<'_>) {
-        if let Ok(line) = ndjson_line(&package_frame(&self.tarball_router, &hint)) {
-            let _ = self.tx.send(line);
-        }
-    }
-
-    fn package_version_guard(&self) -> Option<Arc<dyn PackageVersionGuard>> {
-        self.package_version_guard.clone()
-    }
-}
-
-/// One `package` NDJSON frame. Optional fields are omitted (not null).
-fn package_frame(router: &TarballRouter, hint: &ResolvedPackageHint<'_>) -> serde_json::Value {
-    // A registry-resolved package's `tarball_url` is the packument's
-    // `dist.tarball`, which a split-domain registry hosts on a different origin
-    // — route it by the registry, not the tarball host, so a private package
-    // never leaks its raw upstream URL. Direct tarball deps keep their own URL.
-    let tarball_url = if hint.from_registry {
-        router.route_registry_url(hint.name, hint.version, hint.tarball_url)
-    } else {
-        router.route_url(hint.name, hint.version, hint.tarball_url)
-    };
-    let mut frame = serde_json::json!({
-        "type": "package",
-        "id": hint.id,
-        "name": hint.name,
-        "version": hint.version,
-        "integrity": hint.integrity,
-        "tarball": tarball_url,
-    });
-    if let Some(size) = hint.unpacked_size {
-        frame["unpackedSize"] = serde_json::Value::from(size);
-    }
-    if let Some(count) = hint.file_count {
-        frame["fileCount"] = serde_json::Value::from(count);
-    }
-    if tarball_url == hint.tarball_url
-        && let Some(revision) = hint.revision
-    {
-        frame["revision"] = serde_json::Value::from(revision);
-    }
-    frame
-}
-
-/// `package` frames for every tarball-fetchable entry of a verified
-/// frozen lockfile, deduplicated by tarball URL. Mirrors what the
-/// streaming resolve's [`StreamObserver`] would have announced had the
-/// tree walk run: the client prefetches each tarball on arrival, with
-/// `unpackedSize` (from the verification fan-out's metadata, when the
-/// registry published one) prioritizing the largest downloads.
-///
-/// Tarball URLs are derived with the same
-/// [`tarball_url_and_integrity`] the client's frozen materialization
-/// uses, so the announced URLs match the client's mem-cache keys
-/// byte-for-byte. Non-tarball resolutions (git, directory, binary,
-/// variations) are skipped — the client fetches those through their
-/// own protocol paths.
-fn frozen_package_frames(
-    config: &PacquetConfig,
-    router: &TarballRouter,
-    lockfile: &Lockfile,
-    dist_stats: &ObservedDistStats,
-) -> Vec<Vec<u8>> {
-    let Some(packages) = lockfile.packages.as_ref() else {
-        return Vec::new();
-    };
-    let mut seen_urls = std::collections::HashSet::new();
-    let mut frames = Vec::new();
-    for (package_key, snapshot) in packages {
-        if !matches!(
-            snapshot.resolution,
-            LockfileResolution::Registry(_) | LockfileResolution::Tarball(_),
-        ) {
-            continue;
-        }
-        // The frame carries the integrity the client prefetches against;
-        // an entry that pins none has no frame to announce.
-        let Ok((tarball_url, Some(integrity))) =
-            tarball_url_and_integrity(&snapshot.resolution, package_key, config)
-        else {
-            continue;
-        };
-        let name = package_key.name.to_string();
-        let version = package_key.suffix.version().to_string();
-        let upstream_tarball_url = tarball_url;
-        let tarball_url = router.route_url(&name, &version, &upstream_tarball_url);
-        if !seen_urls.insert(tarball_url.clone()) {
-            continue;
-        }
-        let id = format!("{name}@{version}");
-        let integrity = integrity.to_string();
-        let revision = if tarball_url == upstream_tarball_url {
-            match &snapshot.resolution {
-                LockfileResolution::Tarball(tarball) => tarball.revision.map(TarballRevision::get),
-                LockfileResolution::Registry(registry) => {
-                    registry.revision.map(TarballRevision::get)
-                }
-                _ => None,
-            }
-        } else {
-            None
-        };
-        let stats = dist_stats.get(&(name.clone(), version.clone())).map(|entry| *entry.value());
-        let frame = package_frame(
-            router,
-            &ResolvedPackageHint {
-                id: &id,
-                name: &name,
-                version: &version,
-                integrity: &integrity,
-                tarball_url: &tarball_url,
-                unpacked_size: stats.and_then(|stats| stats.unpacked_size),
-                file_count: stats.and_then(|stats| stats.file_count),
-                revision,
-                // The URL is already routed (canonical → endpoint above), so
-                // re-routing by registry would be redundant; route_url is a
-                // no-op on an already-routed URL.
-                from_registry: false,
-            },
-        );
-        if let Ok(line) = ndjson_line(&frame) {
-            frames.push(line);
-        }
-    }
-    frames
-}
-
-/// Terminal `done` frame: the full resolved lockfile + stats. The client
-/// writes the lockfile and fetches every tarball itself.
-fn done_frame(lockfile: &Lockfile) -> Vec<u8> {
-    let total_packages = lockfile.packages.as_ref().map_or(0, std::collections::HashMap::len);
-    let frame = serde_json::json!({
-        "type": "done",
-        "lockfile": serde_json::to_value(lockfile).unwrap_or(serde_json::Value::Null),
-        "stats": { "totalPackages": total_packages },
-    });
-    ndjson_line(&frame).unwrap_or_else(|_| {
-        br#"{"type":"error","message":"failed to serialize lockfile"}"#.to_vec()
-    })
-}
-
-/// Terminal `error` frame for a resolution that aborted mid-stream,
-/// after one or more `package` frames may already have been sent (so the
-/// HTTP status is locked at 200 — the failure has to ride in the body).
-fn error_frame(message: &str) -> Vec<u8> {
-    let frame = serde_json::json!({ "type": "error", "message": message });
-    ndjson_line(&frame)
-        .unwrap_or_else(|_| br#"{"type":"error","message":"resolution failed"}"#.to_vec())
-}
-
-/// Terminal `violations` frame: the input lockfile failed the client's
-/// policy. Each entry mirrors the local runner's rendered violation so
-/// the client rebuilds the identical `VerifyError` and aborts the same
-/// way the local gate would.
-fn violations_frame(violations: &[serde_json::Value]) -> Vec<u8> {
-    let frame = serde_json::json!({ "type": "violations", "violations": violations });
-    ndjson_line(&frame)
-        .unwrap_or_else(|_| br#"{"type":"error","message":"verification failed"}"#.to_vec())
-}
-
-fn verify_done_frame() -> Vec<u8> {
-    ndjson_line(&serde_json::json!({ "type": "done" }))
-        .unwrap_or_else(|_| br#"{"type":"error","message":"verification failed"}"#.to_vec())
-}
-
-const OSV_VULNERABILITY_CODE: &str = "ERR_PNPM_OSV_VULNERABILITY";
-
-fn verify_done_or_osv_violations(
-    osv_index: Option<&Arc<OsvIndex>>,
-    lockfile: &Lockfile,
-) -> Response {
-    let Some(osv_index) = osv_index else {
-        return ndjson_single_frame(&verify_done_frame());
-    };
-    let violations = osv_violations_for_lockfile(osv_index, lockfile);
-    if violations.is_empty() {
-        ndjson_single_frame(&verify_done_frame())
-    } else {
-        ndjson_single_frame(&violations_frame(&violations))
-    }
-}
-
-fn osv_violations_for_lockfile(index: &OsvIndex, lockfile: &Lockfile) -> Vec<serde_json::Value> {
-    let Some(packages) = lockfile.packages.as_ref() else {
-        return Vec::new();
-    };
-
-    let mut seen = std::collections::HashSet::new();
-    let mut violations = Vec::new();
-    for (package_key, snapshot) in packages {
-        if !is_osv_checkable_resolution(&snapshot.resolution) {
-            continue;
-        }
-        let name = package_key.name.to_string();
-        let version = package_key.suffix.version().to_string();
-        let mut ids = index.vulnerability_ids(&name, &version);
-        // For a tarball resolution the fetched artifact's identity is its
-        // URL, not the lockfile key. Under `trustLockfile` a tampered
-        // lockfile could key a safe `name@version` while pointing the
-        // tarball at a vulnerable artifact, so also screen the version in
-        // the tarball filename. This is additive — a mismatch alone is
-        // never a violation (custom registries may name tarballs
-        // differently), only an actually-vulnerable version is.
-        if let LockfileResolution::Tarball(tarball) = &snapshot.resolution
-            && let Some(url_version) = tarball_url_version(&tarball.tarball, &name)
-            && url_version != version
-        {
-            ids.extend(index.vulnerability_ids(&name, url_version));
-            ids.sort_unstable();
-            ids.dedup();
-        }
-        if ids.is_empty() {
-            continue;
-        }
-        // Dedup only the rare vulnerable hits — several lockfile keys can
-        // share one name@version via peer suffixes — so the common
-        // (non-vulnerable) entry never pays for the set.
-        if !seen.insert((name.clone(), version.clone())) {
-            continue;
-        }
-        violations.push(serde_json::json!({
-            "name": name,
-            "version": version,
-            "code": OSV_VULNERABILITY_CODE,
-            "reason": format!(
-                "is listed in the local OSV database as vulnerable ({})",
-                format_advisory_ids(&ids),
-            ),
-        }));
-    }
-    violations
-}
-
-/// Best-effort extraction of the version from a registry tarball URL of
-/// the conventional `<unscoped-name>-<version>.tgz` shape. Returns `None`
-/// for non-standard naming so a legitimate custom registry isn't
-/// misjudged. Never parses the URL strictly — the lockfile is untrusted.
-fn tarball_url_version<'a>(url: &'a str, name: &str) -> Option<&'a str> {
-    let last = url.rsplit('/').next()?;
-    let last = last.split(['?', '#']).next().unwrap_or(last);
-    let stem = strip_tarball_suffix(last)?;
-    let unscoped = name.rsplit('/').next().unwrap_or(name);
-    let version = stem.strip_prefix(unscoped)?.strip_prefix('-')?;
-    (!version.is_empty()).then_some(version)
-}
-
-/// Strip a `.tgz` / `.tar.gz` tarball suffix case-insensitively, so a
-/// tampered lockfile can't dodge the URL-version cross-check with a
-/// `.TGZ` or `.tar.gz` variant. Returns `None` for any other suffix.
-fn strip_tarball_suffix(name: &str) -> Option<&str> {
-    [".tar.gz", ".tgz"].into_iter().find_map(|suffix| {
-        let head_len = name.len().checked_sub(suffix.len())?;
-        let (head, tail) = (name.get(..head_len)?, name.get(head_len..)?);
-        tail.eq_ignore_ascii_case(suffix).then_some(head)
-    })
-}
-
-fn is_osv_checkable_resolution(resolution: &LockfileResolution) -> bool {
-    match resolution {
-        LockfileResolution::Registry(_) => true,
-        // A frozen lockfile is attacker-controlled, so gate on the tarball
-        // URL rather than the tamper-prone `git_hosted` flag or strict URL
-        // parsing — otherwise `gitHosted: true` or a barely-malformed URL
-        // would let a vulnerable package opt out of the OSV scan. Mirrors
-        // the npm verifier's URL-based gate.
-        LockfileResolution::Tarball(tarball) => {
-            is_http_tarball_url(&tarball.tarball) && !is_git_hosted_tarball_url(&tarball.tarball)
-        }
-        // Custom resolutions are not registry artifacts, so OSV has
-        // no `name@version` advisory coordinates for them.
-        LockfileResolution::Directory(_)
-        | LockfileResolution::Git(_)
-        | LockfileResolution::Binary(_)
-        | LockfileResolution::Variations(_)
-        | LockfileResolution::Custom(_) => false,
-    }
-}
-
-/// Whether a tarball URL uses an http(s) scheme — the only schemes a
-/// registry artifact is served over. Case-insensitive (so a tampered
-/// uppercase scheme can't slip past) without allocating a lowercased copy.
-fn is_http_tarball_url(url: &str) -> bool {
-    let bytes = url.as_bytes();
-    bytes.get(..8).is_some_and(|prefix| prefix.eq_ignore_ascii_case(b"https://"))
-        || bytes.get(..7).is_some_and(|prefix| prefix.eq_ignore_ascii_case(b"http://"))
-}
-
-/// Serialize one frame to a newline-terminated NDJSON line.
-fn ndjson_line(value: &serde_json::Value) -> Result<Vec<u8>, serde_json::Error> {
-    let mut bytes = serde_json::to_vec(value)?;
-    bytes.push(b'\n');
-    Ok(bytes)
-}
-
-/// A 200 NDJSON response carrying a single, already-serialized terminal
-/// frame (the short-circuit and violation paths, which never stream
-/// `package` frames).
-fn ndjson_single_frame(frame: &[u8]) -> Response {
-    Response::builder()
-        .status(StatusCode::OK)
-        .header(header::CONTENT_TYPE, NDJSON_CONTENT_TYPE)
-        .header(PROJECT_TRANSFORMS_HEADER, PROJECT_TRANSFORMS_VERSION)
-        .body(Body::from(frame.to_vec()))
-        .expect("binary response is always valid")
-}
-
-/// A 200 NDJSON response carrying several already-serialized frames in
-/// one fixed body. Used by the frozen fast path, where every frame is
-/// known up front — no channel to stream from.
-fn ndjson_frames(frames: &[Vec<u8>]) -> Response {
-    Response::builder()
-        .status(StatusCode::OK)
-        .header(header::CONTENT_TYPE, NDJSON_CONTENT_TYPE)
-        .header(PROJECT_TRANSFORMS_HEADER, PROJECT_TRANSFORMS_VERSION)
-        .body(Body::from(frames.concat()))
-        .expect("binary response is always valid")
-}
-
-/// A 200 NDJSON response whose body drains the frame channel as the
-/// detached resolve task produces frames. Closing the channel (the task
-/// dropped its sender) ends the body.
-fn ndjson_stream_response(rx: tokio::sync::mpsc::UnboundedReceiver<Vec<u8>>) -> Response {
-    let stream = futures_util::stream::unfold(rx, |mut rx| async move {
-        rx.recv().await.map(|line| (Ok::<_, std::io::Error>(axum::body::Bytes::from(line)), rx))
-    });
-    Response::builder()
-        .status(StatusCode::OK)
-        .header(header::CONTENT_TYPE, NDJSON_CONTENT_TYPE)
-        .header(PROJECT_TRANSFORMS_HEADER, PROJECT_TRANSFORMS_VERSION)
-        .body(Body::from_stream(stream))
-        .expect("streaming response is always valid")
-}
-
 /// Why [`verify_input_lockfile`] failed: either the lockfile violated
 /// the client's policy (carry the rendered violations so the caller can
 /// shape them for the client's protocol) or the verifiers couldn't be
@@ -1481,242 +760,6 @@ fn merge_policies(
         merged.extend(osv_index.policy());
     }
     merged
-}
-
-/// Reject a request that would have pnpr fetch from an origin that is not on
-/// the route allowlist (see [`RouteContext::allows_registry`]) — the
-/// resolver's SSRF boundary, run before any server-side fetch. pnpr fetches
-/// only from operator-configured registries, so a caller cannot point it at
-/// cloud instance metadata, an internal service, or any other off-allowlist
-/// host. Beyond the default/named registries this also covers every fetch a
-/// *direct-URL dependency* would trigger: an `http(s)`/`git` dependency spec,
-/// a catalog or package-extension entry, an override URL leaf, or an input
-/// lockfile's tarball URL. A semver range or `npm:`/`workspace:`/`file:` alias
-/// never hits the network, so it is ignored.
-fn reject_off_allowlist_fetches(
-    request: &ResolveRequest,
-    context: &RouteContext,
-) -> Option<Response> {
-    // Registries are fetch targets whatever their scheme. The registries the
-    // request *declares* are deliberately not checked here: a client describes
-    // its whole configuration, including scopes this resolve never reaches, and
-    // one of those is refused at the fetch instead (`RouteHook::allows_fetch`)
-    // so configuring a registry pnpr does not serve is not itself an error.
-    let mut registries: Vec<&str> = Vec::new();
-    if let Some(registry) = request.registry.as_deref() {
-        registries.push(registry);
-    }
-    if let Some(off) = registries.into_iter().find(|registry| !context.allows_registry(registry)) {
-        return Some(forbidden_off_allowlist(off));
-    }
-
-    // Direct-URL dependency specs and input-lockfile tarball URLs reach the
-    // network only when they carry an http(s)/git URL.
-    let mut url_specs: Vec<&str> = Vec::new();
-    let projects = request.projects_normalized();
-    for project in &projects {
-        for map in
-            [&project.dependencies, &project.dev_dependencies, &project.optional_dependencies]
-        {
-            url_specs.extend(map.values().map(String::as_str));
-        }
-    }
-    if let Some(catalogs) = request.catalogs.as_ref() {
-        url_specs
-            .extend(catalogs.values().flat_map(|catalog| catalog.values()).map(String::as_str));
-    }
-    extend_package_extension_specs(request, &mut url_specs);
-    if let Some(packages) =
-        request.lockfile.as_ref().and_then(|lockfile| lockfile.packages.as_ref())
-    {
-        for package in packages.values() {
-            if let LockfileResolution::Tarball(resolution) = &package.resolution {
-                url_specs.push(resolution.tarball.as_str());
-            }
-        }
-    }
-    if let Some(off) = url_specs.into_iter().find(|spec| fetch_is_off_allowlist(spec, context)) {
-        return Some(forbidden_off_allowlist(off));
-    }
-
-    // Override leaves can themselves be direct-URL specs.
-    if let Some(off) = request
-        .overrides
-        .as_ref()
-        .and_then(|overrides| first_off_allowlist_override(overrides, context))
-    {
-        return Some(forbidden_off_allowlist(&off));
-    }
-
-    None
-}
-
-/// Whether `spec` would trigger a server-side fetch to an origin that is not on
-/// the allowlist. Covers any `scheme://host` URL (an `http(s)` tarball and
-/// every git transport — `git`/`ssh`/`rsync`/`ftp`/`file`/... — with a `git+`
-/// prefix stripped) and scp-style git remotes (`[user@]host:path`), which
-/// pacquet routes to the ssh git resolver. Specs that never reach the network —
-/// semver ranges, `npm:`/`workspace:`/`file:`/`link:` aliases (no `://`),
-/// scoped names — return `false`.
-fn fetch_is_off_allowlist(spec: &str, context: &RouteContext) -> bool {
-    let url = spec.strip_prefix("git+").unwrap_or(spec);
-    if url.contains("://") {
-        // Gate by origin regardless of scheme: any transport that reaches a
-        // host can be an SSRF vector (every git transport — git/ssh/rsync/ftp/
-        // file/...), and a scheme with no allowlistable host (e.g. `file://`,
-        // which would read a server-local path) nerf-darts to nothing and is
-        // rejected.
-        return !context.allows_registry(url);
-    }
-    // A scp-style git remote carries no scheme, so normalize its host to an
-    // `ssh://host/` origin the allowlist can match (nerf-darting is
-    // scheme-agnostic, so an operator allowlisting `https://host/` covers it).
-    match scp_git_host(url) {
-        Some(host) => !context.allows_registry(&format!("ssh://{host}/")),
-        None => false,
-    }
-}
-
-/// The host of a scp-style git remote (`[user@]host:path`), or `None`. The
-/// distinguishing shape is a `user@host` authority before the first `:` with a
-/// path after it — generalizing the `git@...` form pacquet's git resolver treats
-/// as ssh. A protocol spec (`npm:...`, `file:...`) has no `@` in its authority, and
-/// a `scheme://...` URL is handled before this is reached.
-fn scp_git_host(spec: &str) -> Option<&str> {
-    let (authority, path) = spec.split_once(':')?;
-    if path.is_empty() || authority.contains('/') {
-        return None;
-    }
-    let (_, host) = authority.rsplit_once('@')?;
-    (!host.is_empty()).then_some(host)
-}
-
-/// The first override URL leaf whose origin is off the fetch allowlist, if any.
-fn first_off_allowlist_override(
-    value: &serde_json::Value,
-    context: &RouteContext,
-) -> Option<String> {
-    match value {
-        serde_json::Value::String(spec) => {
-            fetch_is_off_allowlist(spec, context).then(|| spec.clone())
-        }
-        serde_json::Value::Array(items) => {
-            items.iter().find_map(|item| first_off_allowlist_override(item, context))
-        }
-        serde_json::Value::Object(map) => {
-            map.values().find_map(|item| first_off_allowlist_override(item, context))
-        }
-        _ => None,
-    }
-}
-
-fn forbidden_off_allowlist(target: &str) -> Response {
-    json_error(
-        StatusCode::FORBIDDEN,
-        &format!(
-            "{target:?} is not allowed by this pnpr server; the operator must declare its \
-             registry as a public route or an upstream",
-        ),
-    )
-}
-
-/// Reject a `registries` map the client could not have loaded itself.
-///
-/// The same validation the config reader runs on the setting, so a request
-/// cannot route one scope to two registries, or reach the resolver with a
-/// declaration pnpm would have refused on disk. The message is the reader's,
-/// with its registry URLs already redacted.
-fn reject_invalid_registries(request: &ResolveRequest) -> Option<Response> {
-    pnpm_config::registries::validate_declarations(request.registries.iter())
-        .err()
-        .map(|error| json_error(StatusCode::BAD_REQUEST, &error.to_string()))
-}
-
-fn reject_invalid_patch_hashes(request: &ResolveRequest) -> Option<Response> {
-    let (selector, _) = request.patched_dependencies.as_ref()?.iter().find(|(_, hash)| {
-        hash.len() != 64
-            || !hash.bytes().all(|byte| byte.is_ascii_hexdigit() && !byte.is_ascii_uppercase())
-    })?;
-    Some(json_error(
-        StatusCode::BAD_REQUEST,
-        &format!("patchedDependencies entry {selector:?} does not contain a SHA-256 hex digest"),
-    ))
-}
-
-/// Reject a request whose client-supplied URLs carry inline
-/// `user:pass@host` credentials, before any fetch or cache write. Covers
-/// the default and named registries, every dependency spec, catalog and
-/// package-extension value, override values, and the tarball URLs of an input
-/// lockfile — every surface a tarball/registry URL can reach the resolver (or
-/// be echoed back) through.
-/// Returns a `400` response when one is found.
-fn reject_inline_url_auth(request: &ResolveRequest) -> Option<Response> {
-    let mut specs: Vec<&str> = Vec::new();
-    if let Some(registry) = request.registry.as_deref() {
-        specs.push(registry);
-    }
-    specs.extend(request.registries.keys().map(String::as_str));
-    let projects = request.projects_normalized();
-    for project in &projects {
-        for map in
-            [&project.dependencies, &project.dev_dependencies, &project.optional_dependencies]
-        {
-            specs.extend(map.values().map(String::as_str));
-        }
-    }
-    if let Some(catalogs) = request.catalogs.as_ref() {
-        specs.extend(catalogs.values().flat_map(|catalog| catalog.values()).map(String::as_str));
-    }
-    extend_package_extension_specs(request, &mut specs);
-    // A supplied lockfile can carry `resolution.tarball` URLs that reach the
-    // verify/frozen paths and would otherwise be routed or echoed back.
-    if let Some(packages) =
-        request.lockfile.as_ref().and_then(|lockfile| lockfile.packages.as_ref())
-    {
-        for package in packages.values() {
-            if let LockfileResolution::Tarball(resolution) = &package.resolution {
-                specs.push(resolution.tarball.as_str());
-            }
-        }
-    }
-    let inline = specs.iter().any(|spec| crate::route::url_has_inline_credentials(spec))
-        || request.overrides.as_ref().is_some_and(overrides_have_inline_url_auth);
-    inline.then(|| {
-        json_error(
-            StatusCode::BAD_REQUEST,
-            "inline URL credentials (user:pass@host) are not allowed; \
-             configure an upstream credential alias instead",
-        )
-    })
-}
-
-fn extend_package_extension_specs<'a>(request: &'a ResolveRequest, specs: &mut Vec<&'a str>) {
-    let Some(extensions) = request.package_extensions.as_ref() else {
-        return;
-    };
-    for extension in extensions.values() {
-        for dependencies in [
-            extension.dependencies.as_ref(),
-            extension.optional_dependencies.as_ref(),
-            extension.peer_dependencies.as_ref(),
-        ]
-        .into_iter()
-        .flatten()
-        {
-            specs.extend(dependencies.values().map(String::as_str));
-        }
-    }
-}
-
-/// Recursively scan an `overrides` JSON value for any string leaf that is
-/// a URL carrying inline credentials.
-fn overrides_have_inline_url_auth(value: &serde_json::Value) -> bool {
-    match value {
-        serde_json::Value::String(spec) => crate::route::url_has_inline_credentials(spec),
-        serde_json::Value::Array(items) => items.iter().any(overrides_have_inline_url_auth),
-        serde_json::Value::Object(map) => map.values().any(overrides_have_inline_url_auth),
-        _ => false,
-    }
 }
 
 fn json_error(status: StatusCode, message: &str) -> Response {
